Improved edit-add designs for game, log, comment
Fixed model for gamelogs, ratings are now decimal.
Updated views for comment sections in game_details.
Added profile user logo in header and a dropdown menu for simplification.

+++ core/models.py
@@ -60,7 +60,7 @@ class GameLog(models.Model):
     game = models.ForeignKey(Game, on_delete=models.CASCADE, related_name="logs")
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='played')
     notes = models.TextField(blank=True, null=True)
-    rating = models.PositiveIntegerField(null=True, blank=True)
+    rating = models.DecimalField(max_digits=3, decimal_places=1, null=True, blank=True)
     hours_played = models.DecimalField(max_digits=5, decimal_places=1, null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
